convert amex and isracard scrapers and fix some types
