+++ src/scrapers/hapoalim.test.js
@@ -3,7 +3,7 @@ import {
   maybeTestCompanyAPI, extendAsyncTimeout, getTestsConfig, exportTransactions,
 } from '../tests/tests-utils';
 import { SCRAPERS } from '../definitions';
-import { LoginResults } from '../constants';
+import { LoginResults } from './base-scraper';
 
 const COMPANY_ID = 'hapoalim'; // TODO this property should be hard-coded in the provider
 const testsConfig = getTestsConfig();
