+++ src/scrapers/otsar-hahayal.js
@@ -1,5 +1,6 @@
 import moment from 'moment';
-import { BaseScraperWithBrowser, LoginResults } from './base-scraper-with-browser';
+import { BaseScraperWithBrowser } from './base-scraper-with-browser';
+import { LoginResults } from './base-scraper';
 import { waitForNavigation } from '../helpers/navigation';
 import {
   fillInput,
